--- Jyros/src/components/Board.css
@@ -1,17 +1,0 @@
-.board-page {
-  display: flex;
-  height: 100vh;
-}
-
-.board {
-    display: flex;
-    justify-content: space-between;
-    gap: 3%;
-  }
-  
-  .board-card {
-    display: flex;
-    flex-direction: column;
-    background-color: lightgrey;
-  }
-
